refactor(commandPalette): ♻️ modularize CommandPaletteListItem
